WeatherModel
The ASMS-B started to collect data since January 2019. Due to some malfunctions of the system, no data were collected for some of the observable nights. To fill these gaps, we have adopted Random Forest Regressions to predict values of seeing, cloud and zenith sky brightness at any time of the year (Lei et al. 2020). 80% of the data collected by ASMS-B in 2019 are adopted as the training sample and the remaining 20% as the testing sample. We have respectively trained the model of the seeing values as a function of the temperature, wind speed and time of the year, the model of the cloudiness as a function of the humid, temperature, wind speed and time of the year, and the model of the zenith sky brightness as a function of the cloudiness and the time of the year. We have compared the values of seeing, cloud and sky brightness predicted by our Random Forest models and those given by the ASMS-B for the test sample. The comparisons show good agreements with no systematic offsets and small dispersions of differences of 0.016′′ , 0.645 and 0.103 mag for seeing, cloudiness and sky brightness, respectively (Lei et al. 2020).